requirements.txt
fastapi
uvicorn[standard]
requests
jinja2
aiofiles
websockets
sqlalchemy
python-dotenv
